+++ aws-transfer/iam-role/variables.tf
@@ -1,0 +1,15 @@
+variable "username" {
+  type = string
+}
+
+variable "s3_bucket_name" {
+  type = string
+}
+
+variable "s3_prefix" {
+  type = string
+}
+
+variable "s3_kms_arn" {
+  type = string
+}